Changed score from float -> integer
Small change to hopefully improve memory/performance, as well as fixing issues with rounding

+++ BoardMoveComparator.java
@@ -4,7 +4,7 @@ public class BoardMoveComparator implements Comparator<BoardMove> {
 
     @Override
     public int compare(BoardMove o1, BoardMove o2) {
-        return Float.compare(o2.board.evaluate(), o1.board.evaluate());
+        return Integer.compare(o2.board.evaluate(), o1.board.evaluate());
     }
 
 }
